+++ graph_coarsening/__init__.py
@@ -1,0 +1,2 @@
+from . import graph_lib, graph_utils
+from .coarsening_utils import coarsen, coarsening_quality, plot_coarsening
